analytics: Rename analytics/views/legacy.py.
