+++ analytics/urls.py
@@ -1,7 +1,7 @@
 from django.conf.urls import include
 from django.urls import path
 
-from analytics.views.legacy import get_activity
+from analytics.views.installation_activity import get_activity
 from analytics.views.realm_activity import get_realm_activity
 from analytics.views.stats import (
     get_chart_data,
