+++ analytics/views/installation_activity.py
